printBody, printStick 함수 추가
