+++ .gitignore
@@ -251,3 +251,4 @@ $RECYCLE.BIN/
 */xcuserdata/*
 
 # End of https://www.toptal.com/developers/gitignore/api/xcode,macos,linux,windows,sublimetext,swift,swiftpackagemanager,swiftpm,objective-c
+*.pbxproj
