--- CodeStarterCamp_Week1/Playground.swift
@@ -1,17 +1,0 @@
-//
-//  Playground.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by 오세용 on 2022/04/16.
-//
-
-import Foundation
-
-func gugu(x: Int) {
-    for count in 1...9 {
-    print(x, " * ", count, " = ", count*x)
-    }
-}
-
-
-// gugu(x:3)
