+++ CodeStarterCamp_Week1/main.swift
@@ -8,30 +8,5 @@
 
 import Foundation
 
-
-func plusTwo(x: Int) -> Int {
-    return x + 2
-}
-
-
-func printBody() {
-    for count in 1...8 {
-        print("**********")
-    }
-}
-
-func printStick() {
-    for count in 1...4 {
-        print("    | |")
-    }
-}
-
-printBody()
-printStick()
-
-gugu(x:3)
-
-/*
- 
- */
-
+printBody(width: 6, height: 5)
+printStick(width: 5, height: 4)
